+++ wholecell/sim/simulation.py
@@ -329,7 +329,7 @@ class Simulation:
             logger.append(self)
 
     def _seedFromName(self, name):
-        return np.uint32((self._seed + hash(name)) % np.iinfo(np.uint64).max)
+        return np.uint32((self._seed + hash(name)) % np.iinfo(np.uint32).max)
         # return np.uint32((self._seed + self.simulationStep() + hash(name)) % np.iinfo(np.uint64).max)
